../../../test/format/util.c:174:16: error: conversion to 'size_t' from 'int' may change the sign of the result [-Werror=sign-conversion]

+++ test/format/util.c
@@ -169,12 +169,12 @@ value_gen(uint8_t *val, uint32_t *sizep, uint64_t keyno)
 char *
 oc_conf(char *buf, size_t buflen, const char *plus)
 {
-	size_t len;
+	int len;
 
 	len = snprintf(buf, buflen, "%s,%s",
 	    plus == NULL ? "" : plus,
 	    DATASOURCE("memrata") ? MEMRATA_DEVICE : "");
-	if (len > buflen)
+	if (len > (int)buflen)
 		die(EINVAL, "open.cursor configuration buffer too small");
 
 	return (buf);
